@@ -9,11 +9,11 @@ import 'package:test/test.dart';
 import 'utils.dart';
 
 void main() {
-  testForImplementation('lazy', ([String string]) {
+  testForImplementation('lazy', ([String? string]) {
     return SpanScanner(string ?? 'foo\nbar\nbaz', sourceUrl: 'source');
   });
 
-  testForImplementation('eager', ([String string]) {
+  testForImplementation('eager', ([String? string]) {
     return SpanScanner.eager(string ?? 'foo\nbar\nbaz', sourceUrl: 'source');
   });
 
@@ -21,7 +21,7 @@ void main() {
     var text = 'first\nbefore: foo\nbar\nbaz :after\nlast';
     var startOffset = text.indexOf('foo');
 
-    SpanScanner scanner;
+    late SpanScanner scanner;
     setUp(() {
       var file = SourceFile.fromString(text, url: 'source');
       scanner =
@@ -49,7 +49,7 @@ void main() {
       scanner.scan('fo');
       scanner.scan('o\nba');
 
-      var span = scanner.lastSpan;
+      var span = scanner.lastSpan!;
       expect(span.start.offset, equals(startOffset + 2));
       expect(span.start.line, equals(1));
       expect(span.start.column, equals(10));
@@ -106,14 +106,14 @@ void main() {
 void testForImplementation(
     String name, SpanScanner Function([String string]) create) {
   group('for a $name scanner', () {
-    SpanScanner scanner;
+    late SpanScanner scanner;
     setUp(() => scanner = create());
 
     test('tracks the span for the last match', () {
       scanner.scan('fo');
       scanner.scan('o\nba');
 
-      var span = scanner.lastSpan;
+      var span = scanner.lastSpan!;
       expect(span.start.offset, equals(2));
       expect(span.start.line, equals(0));
       expect(span.start.column, equals(2));
